update type from size_t to int

+++ search.h
@@ -1,7 +1,7 @@
 #ifndef SEARCH_H
 #define SEARCH_H
 
-int sequential_search(int *array, int k, size_t array_size);
+int sequential_search(int *array, int k, int array_size);
 int sorted_sequential_search(int *array, int k);
 int binary_search(int *array, int k, int array_size);
